Moved spritesheet frames' enums to texture-bank target.
