+++ src/game-loop/include/game-objects/CopyrightsSign.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "MainMenuSpritesheetFrames.hpp"
+#include "spritesheet-frames/MainMenuSpritesheetFrames.hpp"
 #include "components/QuadComponent.hpp"
 #include "GameObject.hpp"
 #include "Point2D.hpp"
@@ -10,7 +10,7 @@ class CopyrightsSign : public GameObject
 public:
 
     CopyrightsSign(float pos_x_center, float pos_y_center)
-            :_quad(TextureType::MAIN_MENU, Renderer::EntityType::MODEL_VIEW_SPACE, COPYRIGHTS_QUAD_WIDTH_WORLD_UNITS, COPYRIGHTS_QUAD_HEIGHT_WORLD_UNITS)
+            :_quad(TextureType::MAIN_MENU, Renderer::EntityType::MODEL_VIEW_SPACE, QUAD_WIDTH_WORLD_UNITS, QUAD_HEIGHT_WORLD_UNITS)
             , _center{pos_x_center, pos_y_center}
     {
         _quad.frame_changed<MainMenuSpritesheetFrames>(MainMenuSpritesheetFrames::COPYRIGHTS);
@@ -23,9 +23,8 @@ public:
 
 private:
 
-    const float COPYRIGHTS_QUAD_WIDTH_WORLD_UNITS = 16.25f;
-    const float COPYRIGHTS_QUAD_HEIGHT_WORLD_UNITS = 1;
-
+    const float QUAD_WIDTH_WORLD_UNITS = 16.25f;
+    const float QUAD_HEIGHT_WORLD_UNITS = 1;
     Point2D _center;
     QuadComponent _quad;
 };
