+++ src/texture-bank/interface/spritesheet-frames/MainDudeSpritesheetFrames.hpp
